feat(blog-article): remove unused code
DA_UK-380

While reviewing the code for the new content page it was difficult to
understand how the blog article page was configured as there was unused
code. While this code may be used in the future, it should not be added
and commented out until needed as this causes confusion.

Also removed unneeded code from the commit lint config and formatted the
code with tslint.